# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Aplikasi REST API",
    "description": "Dokumentasi yang dihasilkan dengan <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/potensidesa/": {
      "get": {
        "tags": [
          "PotensiDesa"
        ],
        "description": "Endpoint untuk mengambil semua PotensiDesa.",
        "responses": {
          "200": {
            "description": "Output API untuk objek PotensiDesa",
            "schema": {
              "$ref": "#/definitions/PotensiDesa"
            }
          }
        }
      },
      "post": {
        "tags": [
          "PotensiDesa"
        ],
        "description": "Endpoint untuk upload file.",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "photo",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "photo cover potensidesa"
          },
          {
            "name": "nama",
            "in": "formData",
            "type": "string",
            "required": true,
            "description": "name potenesidesa"
          },
          {
            "name": "deskrpsi",
            "in": "formData",
            "type": "string",
            "required": true,
            "description": "deskrpsi"
          },
          {
            "name": "alamat",
            "in": "formData",
            "type": "string",
            "required": true,
            "description": "alamat"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nama": {
                  "example": "any"
                },
                "deskrpsi": {
                  "example": "any"
                },
                "alamat": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Profile"
            },
            "description": "Output API untuk objek Profile"
          },
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/potensidesa/{id}": {
      "get": {
        "tags": [
          "PotensiDesa"
        ],
        "description": "Endpoint untuk mengambil PotensiDesa berdasarkan id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek PotensiDesa",
            "schema": {
              "$ref": "#/definitions/PotensiDesa"
            }
          }
        }
      },
      "put": {
        "tags": [
          "PotensiDesa"
        ],
        "description": "Endpoint untuk index http put.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "Informasi PotensiDesa.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PotensiDesa"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek PotensiDesa",
            "schema": {
              "$ref": "#/definitions/PotensiDesa"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PotensiDesa"
        ],
        "description": "Endpoint untuk menghapus PotensiDesa.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek PotensiDesa",
            "schema": {
              "$ref": "#/definitions/PotensiDesa"
            }
          }
        }
      }
    },
    "/statistikJk/": {
      "get": {
        "tags": [
          "StatistikJk"
        ],
        "description": "Endpoint untuk mengambil semua StatistikJk.",
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikJk",
            "schema": {
              "$ref": "#/definitions/StatistikJks"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "StatistikJk"
        ],
        "description": "Endpoint untuk membuat statistikjk baru.",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "Informasi StatistikJk.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StatistikJk"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikJk",
            "schema": {
              "$ref": "#/definitions/StatistikJk"
            }
          }
        }
      }
    },
    "/statistikJk/{id}": {
      "get": {
        "tags": [
          "StatistikJk"
        ],
        "description": "Endpoint untuk mengambil StatistikJk berdasarkan id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikJk",
            "schema": {
              "$ref": "#/definitions/StatistikJk"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "StatistikJk"
        ],
        "description": "Endpoint untuk index http put.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "Informasi StatistikJk.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StatistikJk"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikJk",
            "schema": {
              "$ref": "#/definitions/StatistikJk"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "StatistikJk"
        ],
        "description": "Endpoint untuk menghapus StatistikJk.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikJk",
            "schema": {
              "$ref": "#/definitions/StatistikJk"
            }
          }
        }
      }
    },
...
    "/statistikagama/": {
      "get": {
        "tags": [
          "StatistikAgama"
        ],
        "description": "Endpoint untuk mengambil semua StatistikAgama.",
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikAgama",
            "schema": {
              "$ref": "#/definitions/StatistikAgamas"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "StatistikAgama"
        ],
        "description": "Endpoint untuk membuat statistikagama baru.",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "Informasi StatistikAgama.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StatistikAgama"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikAgama",
            "schema": {
              "$ref": "#/definitions/StatistikAgama"
            }
          }
        }
      }
    },
    "/statistikagama/{id}": {
      "get": {
        "tags": [
          "StatistikAgama"
        ],
        "description": "Endpoint untuk mengambil StatistikAgama berdasarkan id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikAgama",
            "schema": {
              "$ref": "#/definitions/StatistikAgama"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "StatistikAgama"
        ],
        "description": "Endpoint untuk index http put.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "Informasi StatistikAgama.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StatistikAgama"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikAgama",
            "schema": {
              "$ref": "#/definitions/StatistikAgama"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "StatistikAgama"
        ],
        "description": "Endpoint untuk menghapus StatistikAgama.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikAgama",
            "schema": {
              "$ref": "#/definitions/StatistikAgama"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "PotensiDesa": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "situbungur"
        },
        "deskripsi": {
          "type": "string",
          "example": "situbungur"
        },
        "alamat": {
          "type": "string",
          "example": "menjangan"
        },
        "cover": {
          "type": "string",
          "example": "cover.jpg"
        }
      }
    },
...
    "StatistikJk": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "laki_laki": {
          "type": "string",
          "example": "30"
        },
        "perempuan": {
          "type": "string",
          "example": "50"
        }
      }
    },
    "StatistikJks": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/StatistikJk"
      }
    },
...
    "StatistikAgama": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "islam": {
          "type": "string",
          "example": "70"
        },
        "kristen": {
          "type": "string",
          "example": "30"
        },
        "katholik": {
          "type": "string",
          "example": "30"
        },
        "hindu": {
          "type": "string",
          "example": "20"
        },
        "budha": {
          "type": "string",
          "example": "20"
        },
        "konghucu": {
          "type": "string",
          "example": "10"
        },
        "kepercayaan": {
          "type": "string",
          "example": "0"
        }
      }
    },
    "StatistikAgamas": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/StatistikAgama"
      }
    },
...
    "Profile": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "profile-id"
        },
        "name": {
          "type": "string",
          "example": "user 1"
        },
        "email": {
          "type": "string",
          "example": "user1@email.com"
        },
        "occupation": {
          "type": "string",
          "example": "programmer"
        },
        "cover": {
          "type": "string",
          "example": "cover.jpg"
        }
      }
    }
  }
}